+++ packages/blaze-html-templates/package.js
@@ -1,16 +1,16 @@
 Package.describe({
   name: 'blaze-html-templates',
   summary: "Compile HTML templates into reactive UI with Meteor Blaze",
-  version: '2.0.0',
+  version: '3.0.0-rc300.2',
   git: 'https://github.com/meteor/blaze.git'
 });
 
 Package.onUse(function(api) {
   api.imply([
     // A library for reactive user interfaces
-    'blaze@2.7.1',
+    'blaze@3.0.0-rc300.2',
 
     // Compile .html files into Blaze reactive views
-    'templating@1.4.1'
+    'templating@1.4.4-rc300.2'
   ]);
 });