+++ packages/blaze/preamble.js
@@ -8,7 +8,7 @@ Blaze = {};
 // TODO: Should be replaced with _.escape once underscore is upgraded to a newer
 //       version which escapes ` (backtick) as well. Underscore 1.5.2 does not.
 Blaze._escape = (function() {
-  var escape_map = {
+  const escape_map = {
     "<": "&lt;",
     ">": "&gt;",
     '"': "&quot;",
@@ -17,7 +17,7 @@ Blaze._escape = (function() {
     "`": "&#x60;", /* IE allows backtick-delimited attributes?? */
     "&": "&amp;"
   };
-  var escape_one = function(c) {
+  const escape_one = function(c) {
     return escape_map[c];
   };
 
@@ -34,23 +34,19 @@ Blaze._warn = function (msg) {
   }
 };
 
-var nativeBind = Function.prototype.bind;
+const nativeBind = Function.prototype.bind;
 
 // An implementation of _.bind which allows better optimization.
 // See: https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#3-managing-arguments
 if (nativeBind) {
-  Blaze._bind = function (func, obj) {
+  Blaze._bind = function (func, obj, ...rest) {
     if (arguments.length === 2) {
       return nativeBind.call(func, obj);
     }
 
-    // Copy the arguments so this function can be optimized.
-    var args = new Array(arguments.length);
-    for (var i = 0; i < args.length; i++) {
-      args[i] = arguments[i];
-    }
+    const args = [obj, ...rest];
 
-    return nativeBind.apply(func, args.slice(1));
+    return nativeBind.apply(func, args);
   };
 }
 else {