+++ packages/blaze/view_tests.js
@@ -1,11 +1,11 @@
 if (Meteor.isClient) {
 
   Tinytest.add("blaze - view - callbacks", function (test) {
-    var R = ReactiveVar('foo');
+    const R = ReactiveVar('foo');
 
-    var buf = '';
+    let buf = '';
 
-    var v = Blaze.View(function () {
+    const v = Blaze.View(function () {
       return R.get();
     });
 
@@ -24,7 +24,7 @@ if (Meteor.isClient) {
 
     test.equal(buf, '');
 
-    var div = document.createElement("DIV");
+    const div = document.createElement("DIV");
     test.isFalse(v.isRendered);
     test.isFalse(v._isAttached);
     test.equal(canonicalizeHtml(div.innerHTML), "");
@@ -57,10 +57,10 @@ if (Meteor.isClient) {
   });
 
   // this checks, whether a DOMRange is correctly marked as
-  // desroyed after Blaze.remove has destroyed 
+  // desroyed after Blaze.remove has destroyed
   // the corresponding view
   Tinytest.add("blaze - view - destroy", function (test) {
-    var v = {
+    const v = {
       _domrange: Blaze._DOMRange([])
     };
     v._domrange.view = Blaze.View();
@@ -68,11 +68,11 @@ if (Meteor.isClient) {
     Blaze.remove(v);
     test.equal(v._domrange.view.isDestroyed, true);
   });
-  
+
   // this checks, whether an unattached DOMRange notifies
   // correctly about it's root cause, when throwing due to an event
   Tinytest.add("blaze - view - attached", function (test) {
-    test.throws(() => Blaze._DOMRange.prototype.containsElement.call({attached: false, view: {name: 'Template.foo'}}, undefined, '.class', 'click'), 
+    test.throws(() => Blaze._DOMRange.prototype.containsElement.call({attached: false, view: {name: 'Template.foo'}}, undefined, '.class', 'click'),
     `click event triggerd with .class on foo but associated view is not be found.
     Make sure the event doesn't destroy the view.`);
   });