+++ packages/caching-html-compiler/package.js
@@ -2,7 +2,7 @@
 Package.describe({
   name: 'caching-html-compiler',
   summary: 'Pluggable class for compiling HTML into templates',
-  version: '1.2.1',
+  version: '2.0.0-rc300.2',
   git: 'https://github.com/meteor/blaze.git',
 });
 
@@ -12,13 +12,13 @@ Npm.depends({
 
 Package.onUse(function(api) {
   api.use([
-    'caching-compiler@1.2.2',
-    'ecmascript@0.15.1',
+    'caching-compiler@2.0.0-rc300.2',
+    'ecmascript@0.16.9-rc300.2',
   ]);
 
   api.export('CachingHtmlCompiler', 'server');
 
-  api.use(['templating-tools@1.2.1']);
+  api.use(['templating-tools@2.0.0-rc300.2']);
 
   api.addFiles(['caching-html-compiler.js'], 'server');
 });