+++ packages/htmljs/package.js
@@ -1,12 +1,12 @@
 Package.describe({
   name: 'htmljs',
   summary: "Small library for expressing HTML trees",
-  version: '1.1.1',
+  version: '2.0.0-rc300.2',
   git: 'https://github.com/meteor/blaze.git'
 });
 
 Package.onUse(function (api) {
-  api.use('ecmascript@0.15.1');
+  api.use('ecmascript@0.16.9-rc300.2');
 
   api.export('HTML');
   api.mainModule('preamble.js');
@@ -14,7 +14,7 @@ Package.onUse(function (api) {
 
 Package.onTest(function (api) {
   api.use('ecmascript');
-  api.use('tinytest@1.1.0');
+  api.use('tinytest');
 
   api.use('htmljs');