+++ packages/spacebars-tests/template_tests.js
@@ -683,7 +683,7 @@ Tinytest.add('spacebars-tests - template_tests - if in with', function (test) {
 
 Tinytest.add(
   'spacebars-tests - template_tests - each on cursor',
-  function (test) {
+  async function (test) {
     var tmpl = Template.spacebars_template_test_each;
     var coll = new Mongo.Collection(null);
     tmpl.helpers({
@@ -698,15 +698,15 @@ Tinytest.add(
     };
 
     rendersTo('else-clause');
-    coll.insert({ text: 'one', pos: 1 });
+    await coll.insertAsync({ text: 'one', pos: 1 });
     rendersTo('one');
-    coll.insert({ text: 'two', pos: 2 });
+    await coll.insertAsync({ text: 'two', pos: 2 });
     rendersTo('one two');
-    coll.update({ text: 'two' }, { $set: { text: 'three' } });
+    await coll.updateAsync({ text: 'two' }, { $set: { text: 'three' } });
     rendersTo('one three');
-    coll.update({ text: 'three' }, { $set: { pos: 0 } });
+    await coll.updateAsync({ text: 'three' }, { $set: { pos: 0 } });
     rendersTo('three one');
-    coll.remove({});
+    await coll.removeAsync({});
     rendersTo('else-clause');
   }
 );
@@ -795,7 +795,7 @@ Tinytest.add('spacebars-tests - template_tests - ..', function (test) {
   );
 });
 
-Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
+Tinytest.add('spacebars-tests - template_tests - select tags', async function (test) {
   var tmpl = Template.spacebars_template_test_select_tag;
 
   // {label: (string)}
@@ -832,8 +832,8 @@ Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
 
   test.equal(divContent(), ['<select>', '</select>']);
 
-  var optgroup1 = optgroups.insert({ label: 'one' });
-  var optgroup2 = optgroups.insert({ label: 'two' });
+  var optgroup1 = await optgroups.insertAsync({ label: 'one' });
+  var optgroup2 = await optgroups.insertAsync({ label: 'two' });
   test.equal(divContent(), [
     '<select>',
     '<optgroup label="one">',
@@ -843,13 +843,13 @@ Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
     '</select>',
   ]);
 
-  options.insert({
+  await options.insertAsync({
     optgroup: optgroup1,
     value: 'value1',
     selected: false,
     label: 'label1',
   });
-  options.insert({
+  await options.insertAsync({
     optgroup: optgroup1,
     value: 'value2',
     selected: true,
@@ -870,8 +870,8 @@ Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
   test.equal($(selectEl).find('option')[1].selected, true);
 
   // swap selection
-  options.update({ value: 'value1' }, { $set: { selected: true } });
-  options.update({ value: 'value2' }, { $set: { selected: false } });
+  await options.updateAsync({ value: 'value2' }, { $set: { selected: false } });
+  await options.updateAsync({ value: 'value1' }, { $set: { selected: true } });
   Tracker.flush();
 
   test.equal(divContent(), [
@@ -889,8 +889,8 @@ Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
   test.equal($(selectEl).find('option')[1].selected, false);
 
   // change value and label
-  options.update({ value: 'value1' }, { $set: { value: 'value1.0' } });
-  options.update({ value: 'value2' }, { $set: { label: 'label2.0' } });
+  await options.updateAsync({ value: 'value1' }, { $set: { value: 'value1.0' } });
+  await options.updateAsync({ value: 'value2' }, { $set: { label: 'label2.0' } });
   Tracker.flush();
 
   test.equal(divContent(), [
@@ -910,17 +910,17 @@ Tinytest.add('spacebars-tests - template_tests - select tags', function (test) {
   // unselect and then select both options. normally, the second is
   // selected (since it got selected later). then switch to <select
   // multiple="">. both should be selected.
-  options.update({}, { $set: { selected: false } }, { multi: true });
+  await options.updateAsync({}, { $set: { selected: false } }, { multi: true });
   Tracker.flush();
-  options.update({}, { $set: { selected: true } }, { multi: true });
+  await options.updateAsync({}, { $set: { selected: true } }, { multi: true });
   Tracker.flush();
   test.equal($(selectEl).find('option')[0].selected, false);
   test.equal($(selectEl).find('option')[1].selected, true);
 
   selectEl.multiple = true; // allow multiple selection
-  options.update({}, { $set: { selected: false } }, { multi: true });
+  await options.updateAsync({}, { $set: { selected: false } }, { multi: true });
   Tracker.flush();
-  options.update({}, { $set: { selected: true } }, { multi: true });
+  await options.updateAsync({}, { $set: { selected: true } }, { multi: true });
   Tracker.flush();
   test.equal($(selectEl).find('option')[0].selected, true);
   test.equal($(selectEl).find('option')[1].selected, true);
@@ -1542,6 +1542,54 @@ Tinytest.add(
     dep.changed();
     Tracker.flush();
     test.equal(pElement.getAttribute('foo'), 'not-bar');
+  }
+);
+
+// The attribute object could be disabled or null, which
+// should be handled, as if an empty object is passed
+Tinytest.add(
+  'spacebars-tests - template_tests - attribute object helpers are disabled',
+  function (test) {
+    const tmpl =
+      Template.spacebars_template_test_attr_object_helpers_are_disabled;
+    tmpl.helpers({
+      disabled: function () {
+        return undefined;
+      },
+    });
+
+    // should not throw
+    const div = renderToDiv(tmpl);
+
+    // button should not be affected
+    const pElement = div.querySelector('button');
+    test.equal(pElement.getAttribute('title'), null);
+    const text = pElement.firstChild.textContent;
+    test.equal(text, 'test');
+  }
+);
+
+// The attribute object could be disabled or null, which
+// should be handled, as if an empty object is passed
+Tinytest.add(
+  'spacebars-tests - template_tests - attribute object helpers are disabled should not affect existing atts',
+  function (test) {
+    const tmpl =
+      Template.spacebars_template_test_attr_object_helpers_are_disabled2;
+    tmpl.helpers({
+      disabled: function () {
+        return undefined;
+      },
+    });
+
+    // should not throw
+    const div = renderToDiv(tmpl);
+
+    // existing atts should not be affected
+    const pElement = div.querySelector('button');
+    test.equal(pElement.getAttribute('title'), 'foo');
+    const text = pElement.firstChild.textContent
+    test.equal(text, 'test');
   }
 );
 
@@ -2782,7 +2830,7 @@ Tinytest.add(
 // https://github.com/meteor/meteor/issues/2156
 Tinytest.add(
   'spacebars-tests - template_tests - each with inserts inside autorun',
-  function (test) {
+  async function (test) {
     var tmpl = Template.spacebars_test_each_with_autorun_insert;
     var coll = new Mongo.Collection(null);
     var rv = new ReactiveVar();
@@ -2810,7 +2858,7 @@ Tinytest.add(
 
     test.equal(canonicalizeHtml(div.innerHTML), 'foo1 foo2');
 
-    coll.update(firstId, { $set: { name: 'foo3' } });
+    await coll.updateAsync(firstId, { $set: { name: 'foo3' } });
     Tracker.flush();
     test.equal(canonicalizeHtml(div.innerHTML), 'foo3 foo2');
   }
@@ -4085,12 +4133,12 @@ Tinytest.add(
 
 Tinytest.add(
   'spacebars-tests - template_tests - #each @index',
-  function (test) {
+  async function (test) {
     var tmpl = Template.spacebars_template_test_each_index;
 
     var c = new Mongo.Collection();
-    c.insert({ num: 2 });
-    c.insert({ num: 4 });
+    await c.insertAsync({ num: 2 });
+    await c.insertAsync({ num: 4 });
     tmpl.helpers({
       things: function () {
         return c.find({}, { sort: { num: 1 } });
@@ -4104,28 +4152,28 @@ Tinytest.add(
       '<span>0 - 2</span><span>1 - 4</span>'
     );
 
-    c.insert({ num: 1 });
+    await c.insertAsync({ num: 1 });
     Tracker.flush();
     test.equal(
       canonicalizeHtml(div.innerHTML),
       '<span>0 - 1</span><span>1 - 2</span><span>2 - 4</span>'
     );
 
-    var three = c.insert({ num: 3 });
+    var three = await c.insertAsync({ num: 3 });
     Tracker.flush();
     test.equal(
       canonicalizeHtml(div.innerHTML),
       '<span>0 - 1</span><span>1 - 2</span><span>2 - 3</span><span>3 - 4</span>'
     );
 
-    c.update(three, { num: 0 });
+    await c.updateAsync(three, { num: 0 });
     Tracker.flush();
     test.equal(
       canonicalizeHtml(div.innerHTML),
       '<span>0 - 0</span><span>1 - 1</span><span>2 - 2</span><span>3 - 4</span>'
     );
 
-    c.remove(three);
+    await c.removeAsync(three);
     Tracker.flush();
     test.equal(
       canonicalizeHtml(div.innerHTML),