+++ packages/spacebars-tests/template_tests_server.js
@@ -2,7 +2,7 @@ const path = Npm.require("path");
 
 Meteor.methods({
   getAsset: function (filename) {
-    return Assets.getText(path.join("assets", filename));
+    return Assets.getTextAsync(path.join("assets", filename));
   }
 });