+++ packages/ui/package.js
@@ -1,13 +1,13 @@
 Package.describe({
   name: 'ui',
   summary: "Deprecated: Use the 'blaze' package",
-  version: '1.0.14',
+  version: '2.0.0-rc300.2',
   git: 'https://github.com/meteor/blaze.git',
   deprecated: true,
 });
 
 Package.onUse(function (api) {
-  api.use('blaze@2.3.2');
+  api.use('blaze@3.0.0-rc300.2');
   api.imply('blaze');
 
   // XXX COMPAT WITH PACKAGES BUILT FOR 0.9.0.